requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==23.1.2
Automat==24.8.1
backports.zoneinfo==0.2.1
cachecontrol==0.14.0
cachetools==5.5.0
certifi==2022.12.7
cffi==1.17.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.0
daphne==4.1.2
dj-database-url==2.2.0
Django==4.2.13
django-admin-geomap==1.6
django-cors-headers==4.4.0
django-jazzmin==3.0.0
django-oauth-toolkit==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.7
fcm-django==2.2.1
firebase-admin==6.5.0
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
gunicorn==23.0.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
inflection==0.5.1
itypes==1.2.0
jinja2==3.1.4
jwcrypto==1.5.6
MarkupSafe==2.1.5
msgpack==1.1.0
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
service-identity==24.1.0
simplejson==3.19.3
sqlparse==0.5.1
swapper==1.4.0
tomli==2.0.1
twisted==24.7.0
txaio==23.1.1
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.7.0
zope.interface==7.0.3
